simplify code and fix ci
