+++ src/iceberg/test/CMakeLists.txt
@@ -97,7 +97,6 @@ add_iceberg_test(logging_test
                  log_level_test.cc
                  logger_test.cc
                  loggers_test.cc
-                 logging_end_to_end_test.cc
                  macros_active_level_test.cc
                  macros_test.cc
                  spdlog_logger_test.cc)
